Skip XERCES cegui dependency compilation on mingw
intrin.h doesn't exist on standard MinGW distribution.
We also import the mingw-w64 config header, even if we're working with
the standard (32) mingw version. That might need further investigation, but
either way, normal mingw 32 doesn't work with xerces. We might want to enable
mingw-w64 support for the builder in the future